refactor!: game betting and action rewrite
